+++ algojkl-frontend/src/App.css
@@ -9,7 +9,7 @@ body {
   width: 100%;
 }
 .starter {
-  width: 100%;
+  width: 100% !important;
 }
 
 /* -------------------------------------------Carousel --------------------------------------------------------*/
@@ -742,8 +742,11 @@ body.modal-open {
   padding: 10px;
 }
 
-.Lakki-container img {
-  width: 100%;
+.lakki {
+  width: 73%;
+  padding: 10px;
+  margin: 0 auto;
+  display: block;
 }
 /* ------------------------------------------------------------------------------------------------ */
