+++ algojkl-frontend/src/components/PestitDescription.jsx
@@ -1,0 +1,75 @@
+import React from 'react'
+
+const PestitDescription = () => (
+    <div className="hallitus-pesti-info">
+    <h3>Lyhyet kuvaukset hallituspesteistä</h3>
+    <p>
+      <strong>Puheenjohtaja</strong> on vastuussa hallituksen toiminnasta,
+      johtaa hallituksen kokouksia sekä edustaa yleisesti kiltaa.
+      <br />
+      <br />
+      <strong>Varapuheenjohtaja</strong> sijaistaa puheenjohtajaa
+      kokouksissa tämän estyessä, ja toimii tiiviissä yhteistyössä
+      puheenjohtajan kanssa.
+      <br />
+      <br />
+      <strong>Sihteeri</strong> kirjoittaa kokouksissa pöytäkirjan
+      (virallinen dokumentti kokouksissa läpikäytävistä asioista).
+      <br />
+      <br />
+      <strong>Rahastonhoitaja</strong> maksaa laskuja, huolehtii
+      kirjanpidosta ja vastaa killan pankkitilistä.
+      <br />
+      <br />
+      <strong>Yritysvastaava</strong> solmii ja ylläpitää
+      yhteistyösopimuksia yritysten kanssa sekä vastaa yritystiimin
+      toiminnasta.
+      <br />
+      <br />
+      <strong>Tapahtumavastaava</strong> ideoi ja järjestää tapahtumia sekä
+      vastaa tapahtumatiimin toiminnasta.
+      <br />
+      <br />
+      <strong>Projektivastaavat</strong> vastaavat teekkariuteen liittyvistä
+      projektiluontoisista tapahtumista kuten vuosijuhlista,
+      lakinlaskijaisista ja kastajaisista.
+      <br />
+      <br />
+      <strong>Koulutuspoliittinen vastaava</strong> (kopo) huolehtii
+      opiskelijoiden oikeuksien toteutumisesta ja ylläpitää suhteita
+      yliopistoon.
+      <br />
+      <br />
+      <strong>Sosiaalipoliittinen vastaava</strong> (sopo) huolehtii
+      opiskelijoiden edunvalvonnasta, hyvinvoinnista ja yhdenvertaisuudesta
+      killan toiminnassa.
+      <br />
+      <br />
+      <strong>Kerhovastaava</strong> vastaa kommunikaatiosta kerhojen ja
+      hallituksen välillä
+      <br />
+      <br />
+      <strong>Viestintävastaava</strong> ylläpitää killan sosiaalisen median
+      tilejä ja tiedottaa jäsenille tapahtumista ja muista tärkeistä
+      asioista.
+      <br />
+      <br />
+      <br />
+      <strong></strong>
+      <i>
+        ps: Hallituksessa tehdään yhteistyössä muitakin kuin vain oman
+        vastuualueen tehtäviä.
+      </i>
+    </p>
+    <br />
+    <p>
+      {' '}
+      Aiemmat hallitukset näet{' '}
+      <a href="/entiset-hallitukset">
+        <strong>täältä</strong>
+      </a>
+    </p>
+  </div>
+)
+
+export default PestitDescription
