+++ algojkl-frontend/src/utils/hallitusOrder.jsx
@@ -1,0 +1,9 @@
+export const hallitusOrder = {
+    'Puheenjohtaja, President': 1,
+    'Varapuheenjohtaja, Vice president': 2,
+    'Rahastonhoitaja, Treasurer': 3,
+    'Sihteeri, Secretary': 4,
+    'Koulutuspoliittinen vastaava, Head of educational politics, (kopo)': 5,
+    'Sosiaalipolittinen vastaava, Head of social affairs, (sopo)': 6,
+  }
+
